refactor: Modularizes the js of assets
