--- public/js/assets/assets.js
@@ -1,602 +1,0 @@
-// ------------------------------
-// Modal dinámico por categoría
-// ------------------------------
-const categoriaSelect = document.getElementById('categoria');
-const camposDinamicos = document.getElementById('camposDinamicos');
-const csrfToken = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
-let editingAssetId = null;
-let assetIdToDelete = null;
-let categoriasGlobales = [];
-
-const camposGenericos = `
-    <h6 class="text-uppercase text-secondary fw-semibold mb-3">
-        <i class="fas fa-cogs me-2"></i>Detalles Específicos
-    </h6>
-    <div class="row g-3 mt-2">
-        <div class="col-md-6">
-            <div class="form-floating">
-                <input type="text" class="form-control" id="procesador" placeholder="Procesador">
-                <label for="procesador">Procesador</label>
-            </div>
-        </div>
-        <div class="col-md-3">
-            <div class="form-floating">
-                <input type="text" class="form-control" id="velocidad" placeholder="Velocidad">
-                <label for="velocidad">Velocidad</label>
-            </div>
-        </div>
-        <div class="col-md-3">
-            <div class="form-floating">
-                <input type="text" class="form-control" id="memoria" placeholder="Memoria">
-                <label for="memoria">Memoria</label>
-            </div>
-        </div>
-        <div class="col-md-12">
-            <div class="form-floating">
-                <input type="text" class="form-control" id="almacenamiento" placeholder="Capacidad de almacenamiento">
-                <label for="almacenamiento">Capacidad de almacenamiento</label>
-            </div>
-        </div>
-    </div>
-`;
-
-// ------------------------------
-// Detecta la categoria 
-// ------------------------------
-if (categoriaSelect) {
-    categoriaSelect.addEventListener('change', () => {
-        const categoriaId = parseInt(categoriaSelect.value);
-        camposDinamicos.innerHTML = '';
-
-        if (!categoriaId) return;
-
-        const categoriaSeleccionada = categoriasGlobales.find(cat => cat.id === categoriaId);
-        if (categoriaSeleccionada && categoriaSeleccionada.special_specifications) {
-            camposDinamicos.innerHTML = camposGenericos;
-        }
-    });
-}
-
-// ------------------------------
-// Cargar categorías dinámicamente
-// ------------------------------
-async function cargarCategorias(selectedId = null) {
-    
-    const categoriaSelect = document.getElementById('categoria');
-    if (!categoriaSelect) {
-        return;
-    }
-
-    try {
-        const res = await fetch('/categories/api', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                'X-CSRF-TOKEN': csrfToken
-            }
-        });
-
-        const json = await res.json();
-
-        if (!json.ok) {
-            showAlert(json.message || 'Error al cargar las categorías', "red", "Error");
-            return;
-        }
-
-        categoriasGlobales = json.data;
-
-        categoriaSelect.innerHTML = `<option value="" selected>Seleccione categoría</option>`;
-
-        let optionsCount = 0;
-        json.data.forEach(cat => {
-            const option = document.createElement('option');
-            option.value = cat.id;
-            option.textContent = cat.name;
-            if (selectedId && selectedId == cat.id) {
-                option.selected = true;
-            }
-            categoriaSelect.appendChild(option);
-            optionsCount++;
-        });
-
-    } catch (error) {
-        console.error('Error en cargarCategorias:', error);
-        showAlert('Error en la carga de categorías: ' + error.message, "red", "Error");
-    }
-}
-
-// ------------------------------
-// Cargar marcas dinámicamente
-// ------------------------------
-async function cargarMarcas(selectedId = null) {
-    
-    try {
-        const marcaSelect = document.getElementById('marca');
-        if (!marcaSelect) {
-            return;
-        }
-
-        const res = await fetch('/brands/api', {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                'X-CSRF-TOKEN': csrfToken
-            }
-        });
-
-        const json = await res.json();
-
-        if (!json.ok) {
-            console.error('Error en respuesta JSON:', json.message);
-            showAlert(json.message || 'Error al cargar marcas', "red", "Error");
-            return;
-        }
-
-        marcaSelect.innerHTML = '<option value="">Seleccione marca</option>';
-
-        let optionsCount = 0;
-        json.data.forEach(m => {
-            const option = document.createElement('option');
-            option.value = m.id;
-            option.textContent = m.name;
-            if (selectedId && selectedId == m.id) {
-                option.selected = true;
-            }
-            marcaSelect.appendChild(option);
-            optionsCount++;
-        });
-
-    } catch (error) {
-        console.error('Error en cargarMarcas:', error);
-        showAlert('Error al cargar las marcas: ' + error.message, "red", "Error");
-    }
-}
-
-// ------------------------------
-// DOMContentLoaded
-// ------------------------------
-document.addEventListener('DOMContentLoaded', function () {
-
-    // ------------------------------
-    // Inicialización de DataTable
-    // ------------------------------
-    fetch('/assets/api', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-            'X-CSRF-TOKEN': csrfToken
-        },
-        body: JSON.stringify({ option: 'table' }) 
-    })
-    .then(response => response.json())
-    .then(json => {
-        if (!json.ok) {
-            showAlert('Error al cargar los activos: ' + (json.message || 'Desconocido'), "red", "Error");
-            return;
-        }
-
-        const columns = [
-            { data: 'inventory_number', className: 'text-center fw-medium', title: 'N° Inventario' },
-            { data: 'model', className: 'fw-normal', title: 'Modelo' },
-            { data: 'serial_number', className: 'fw-normal', title: 'N° Serie' },
-            { data: 'brand.name', className: 'fw-normal', title: 'Marca' },
-            { data: 'category.name', className: 'fw-normal', title: 'Categoría' },
-            {
-                data: 'is_active_label',
-                className: 'text-center',
-                title: 'Estado',
-                render: function (data, type, row) {
-                    const badgeClass = row.is_active ? 'bg-success' : 'bg-danger';
-                    return `<span class="badge ${badgeClass} rounded-pill px-3 py-1">${data}</span>`;
-                }
-            },
-            {
-                data: null,
-                className: 'text-center',
-                orderable: false,
-                title: 'Acciones',
-                render: (data, type, row) => `
-                    <button
-                        class="btn btn-outline-info btn-sm mx-1 btn-ver"
-                        title="Ver"
-                        data-id="${row.id}">
-                        <i class="fas fa-eye"></i>
-                    </button>
-                    <button class="btn btn-outline-primary btn-sm mx-1 btn-modal-bien"
-                        data-mode="edit"
-                        data-id="${row.id}">
-                        <i class="fas fa-edit"></i>
-                    </button>
-                    <button
-                        class="btn btn-outline-danger btn-sm mx-1 btn-delete-asset"
-                        data-id="${row.id}"
-                        title="Eliminar">
-                        <i class="fas fa-trash"></i>
-                    </button>
-                `
-            }
-        ];
-
-        bottomTableConfig('file_export', json.data, columns, '[title]');
-    })
-    .catch(error => {
-        showAlert('Error al cargar los activos: ' + error.message, "red", "Error de conexión");
-    });
-
-    // ------------------------------
-    // Variables del modal
-    // ------------------------------
-    const modalBien = document.getElementById('modalBien');
-    const modalTitle = document.getElementById('modalBienTitulo');
-    const formBien = document.getElementById('formNuevoBien');
-    const btnSubmit = modalBien ? modalBien.querySelector('button[type="submit"]') : null;
-    const tableBody = document.querySelector('#file_export tbody');
-
-    if (!modalBien) {
-        return;
-    }
-
-    // ------------------------------
-    // CONFIGURACIÓN DE BOTONES DE CIERRE DEL MODAL
-    // ------------------------------
-
-    // Botón de cerrar en el header del modal (la X)
-    closeModal('btnCerrarModalBien', 'modalBien', 'focusAfterSave');
-
-    // Botón "Cerrar" en el footer del modal  
-    closeModal('btnCerrarFooter', 'modalBien', 'focusAfterSave');
-
-    // ------------------------------
-    // Cargar tabla
-    // ------------------------------
-    function loadAssets() {
-        if ($.fn.DataTable.isDataTable('#file_export')) {
-            const table = $('#file_export').DataTable();
-
-            fetch('/assets/api', {  
-                method: 'POST',   
-                headers: {
-                    'Content-Type': 'application/json',
-                    'X-CSRF-TOKEN': csrfToken
-                },
-                body: JSON.stringify({ 
-                    option: 'table' 
-                })
-            })
-            .then(res => res.json())
-            .then(json => {
-                if (!json.ok) return;
-                table.clear();
-                table.rows.add(json.data);
-                table.draw();
-            })
-            .catch(err => {
-                console.error("Error al refrescar la tabla:", err);
-                showAlert(
-                    "Error de conexión al intentar refrescar la tabla.",
-                    "red",
-                    "Error loadAssets",
-                    () => console.log("Error al refrescar la tabla:", err),
-                    3000
-                );
-            });
-        }
-    }
-
-    // Carga inicial
-    loadAssets();
-
-    // ------------------------------
-    // Abrir modal según modo
-    // ------------------------------
-    document.addEventListener('click', async e => {
-        const btn = e.target.closest('.btn-modal-bien');
-        if (!btn) return;
-
-        const mode = btn.dataset.mode;
-        const id = btn.dataset.id || null;
-        
-        if (formBien) formBien.reset();
-
-        if (mode === 'create') {
-            modalTitle.textContent = 'Nuevo Bien';
-            btnSubmit.innerHTML = '<i class="fas fa-plus me-1"></i> Guardar';
-            btnSubmit.classList.remove('btn-warning');
-            btnSubmit.classList.add('btn-primary');
-            formBien.dataset.mode = 'create';
-            delete formBien.dataset.id;
-
-            try {
-                await Promise.all([
-                    cargarCategorias(), 
-                    cargarMarcas() 
-                ]);
-                openModalForEdit('modalBien');
-            } catch (error) {
-                console.error('Error en Promise.all:', error);
-                showAlert('Error al cargar los datos del formulario', 'red', 'Error');
-                return;
-            }
-        }
-
-        if (mode === 'edit') {
-
-            modalTitle.textContent = 'Editar Bien';
-            btnSubmit.innerHTML = '<i class="fas fa-save me-1"></i> Actualizar';
-            btnSubmit.classList.remove('btn-primary');
-            btnSubmit.classList.add('btn-warning');
-            formBien.dataset.mode = 'edit';
-            formBien.dataset.id = id;
-
-            fetch('/assets/api', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'X-CSRF-TOKEN': csrfToken
-                },
-                body: JSON.stringify({ 
-                    option: 'details',
-                    id: id 
-                })
-            })
-            .then(res => {
-                return res.json();
-            })
-            .then(async result => {
-
-                if (!result.ok) {
-                    console.error('Error en respuesta de details');
-                    return;
-                }
-                const asset = Array.isArray(result.data) 
-                    ? result.data.find(a => a.id == id)
-                    : result.data;
-                    
-                if (!asset) {
-                    console.error('Asset no encontrado');
-                    return;
-                }
-
-                document.getElementById('numeroInventario').value = asset.inventory_number ?? '';
-                document.getElementById('marca').value = asset.brand_id ?? '';
-                document.getElementById('modelo').value = asset.model ?? '';
-                document.getElementById('serie').value = asset.serial_number ?? '';
-                document.getElementById('estado').value = asset.is_active ? '1' : '0';
-                document.getElementById('categoria').value = asset.category_id ?? '';
-                document.getElementById('descripcion').value = asset.description ?? '';
-
-                try {
-                    await Promise.all([
-                        cargarCategorias(asset.category_id), 
-                        cargarMarcas(asset.brand_id)
-                    ]);
-                } catch (error) {
-                    console.error('Error cargando datos:', error);
-                    showAlert('Error al cargar los datos del formulario', 'red', 'Error');
-                    return;
-                }
-
-                const categoriaSeleccionada = categoriasGlobales.find(cat => cat.id === asset.category_id);
-                if (categoriaSeleccionada && categoriaSeleccionada.special_specifications) {
-                    camposDinamicos.innerHTML = camposGenericos;
-                } else {
-                    camposDinamicos.innerHTML = '';
-                }
-
-                if (document.getElementById('procesador')) document.getElementById('procesador').value = asset.cpu ?? '';
-                if (document.getElementById('velocidad')) document.getElementById('velocidad').value = asset.speed ?? '';
-                if (document.getElementById('memoria')) document.getElementById('memoria').value = asset.memory ?? '';
-                if (document.getElementById('almacenamiento')) document.getElementById('almacenamiento').value = asset.storage ?? '';
-
-                openModalForEdit('modalBien');
-            })
-            .catch(err => {
-                console.error('Error en fetch details:', err);
-                showAlert(
-                    "Error al obtener datos del fetch.",
-                    "red",
-                    "Error Fetch",
-                    () => console.log("Error al cargar los datos"),
-                    3000
-                )
-            });
-        }
-    });
-
-    // ------------------------------
-    // Envío del formulario
-    // ------------------------------
-    if (formBien) {
-        formBien.addEventListener('submit', async e => {
-            e.preventDefault();
-            const mode = formBien.dataset.mode;
-
-            const formData = {
-                inventory_number: document.getElementById('numeroInventario').value,
-                brand_id: document.getElementById('marca').value,
-                model: document.getElementById('modelo').value,
-                serial_number: document.getElementById('serie').value,
-                is_active: document.getElementById('estado').value === '1',
-                category_id: document.getElementById('categoria').value,
-                description: document.getElementById('descripcion').value,
-                cpu: document.getElementById('procesador')?.value || null,
-                speed: document.getElementById('velocidad')?.value || null,
-                memory: document.getElementById('memoria')?.value || null,
-                storage: document.getElementById('almacenamiento')?.value || null
-            };
-
-            const action = async () => {
-                try {
-                    let res;
-
-                    if (mode === 'create') {
-                        res = await fetch('/assets', {
-                            method: 'POST',
-                            headers: { 'Content-Type': 'application/json', 'X-CSRF-TOKEN': csrfToken },
-                            body: JSON.stringify(formData)
-                        });
-                    } else {
-                        const id = formBien.dataset.id;
-                        res = await fetch(`/assets/${id}`, {
-                            method: 'PUT',
-                            headers: { 'Content-Type': 'application/json', 'X-CSRF-TOKEN': csrfToken },
-                            body: JSON.stringify(formData)
-                        });
-                    }
-
-                    const result = await res.json();
-
-                    if (!result.ok) {
-                        if (res.status === 422 && result.errors) {
-                            let messages = Object.values(result.errors)
-                                .flat()
-                                .join('<br>');
-                            showAlert(messages, "red", "Error de validación");
-                            return;
-                        }
-
-                        showAlert(result.message || 'Error en la operación.', "red", "Error");
-                        return;
-                    }
-
-                    showAlert(result.message || 'Operación exitosa.', "green", "Éxito", () => {
-                        closeModalForSuccess('modalBien', 'focusAfterSave');
-                        loadAssets();
-                    });
-
-                } catch (error) {
-                    showAlert('Error al enviar el formulario: ' + error.message, "red", "Error");
-                }
-            };
-
-            if (mode === 'create') confirmStore(action);
-            else confirmUpdate(action);
-        });
-    }
-
-    // ------------------------------
-    // Transforma a mayúscula todos los inputs
-    // ------------------------------
-    if (formBien) {
-        formBien.addEventListener('input', e => {
-            const target = e.target;
-            if ((target.tagName === 'INPUT' && target.type === 'text') || target.tagName === 'TEXTAREA') {
-                const start = target.selectionStart;
-                const end = target.selectionEnd;
-                target.value = target.value.toUpperCase();
-                target.setSelectionRange(start, end);
-            }
-        });
-    }
-
-    // ------------------------------
-    // Evento delegado: Ver detalles
-    // ------------------------------
-    document.querySelector('#file_export tbody').addEventListener('click', async (event) => {
-        const button = event.target.closest('.btn-ver');
-        if (!button) return;
-
-        const id = button.dataset.id;
-        if (!id) return;
-
-        try {
-            const response = await fetch('/assets/api', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'X-CSRF-TOKEN': csrfToken
-                },
-                body: JSON.stringify({ option: 'details' })
-            });
-            const result = await response.json();
-            if (!result.ok) {
-                showAlert(result.message || 'Error al cargar detalles del activo.', "red", "Error");
-                return;
-            }
-
-            const asset = result.data.find(a => a.id == id);
-            if (!asset) {
-                showAlert('No se encontró la información del activo.', "orange", "Advertencia");
-                return;
-            }
-
-            const modal = document.getElementById('modalDetallesBien');
-
-            function formatDate(isoString) {
-                if (!isoString) return '—';
-                const date = new Date(isoString);
-                return date.toLocaleString('es-MX', {
-                    year: 'numeric',
-                    month: '2-digit',
-                    day: '2-digit',
-                    hour: '2-digit',
-                    minute: '2-digit',
-                });
-            }
-
-            modal.querySelectorAll('span[id^="detalle-"]').forEach(span => span.textContent = '...');
-            modal.querySelector('#detalle-descripcion').textContent = '';
-
-            modal.querySelector('#detalle-inventario').textContent = asset.inventory_number ?? '—';
-            modal.querySelector('#detalle-modelo').textContent = asset.model ?? '—';
-            modal.querySelector('#detalle-serie').textContent = asset.serial_number ?? '—';
-            modal.querySelector('#detalle-marca').textContent = asset.brand?.name ?? '—';
-            modal.querySelector('#detalle-categoria').textContent = asset.category?.name ?? '—';
-
-            const estadoSpan = modal.querySelector('#detalle-estado');
-            estadoSpan.textContent = asset.is_active_label;
-            estadoSpan.classList.remove('bg-success', 'bg-danger');
-            estadoSpan.classList.add(asset.is_active ? 'bg-success' : 'bg-danger');
-
-            modal.querySelector('#detalle-creado').textContent = formatDate(asset.created_at);
-            modal.querySelector('#detalle-cpu').textContent = asset.cpu ?? '—';
-            modal.querySelector('#detalle-velocidad').textContent = asset.speed ?? '—';
-            modal.querySelector('#detalle-memoria').textContent = asset.memory ?? '—';
-            modal.querySelector('#detalle-almacenamiento').textContent = asset.storage ?? '—';
-            modal.querySelector('#detalle-descripcion').textContent = asset.description ?? '—';
-
-            openModalForEdit('modalDetallesBien');
-        } catch (error) {
-            showAlert('Error al cargar los detalles: ' + error.message, "red", "Error");
-        }
-    });
-
-    // ------------------------------
-    // Eliminar bien
-    // ------------------------------
-    document.addEventListener('click', (e) => {
-        const deleteButton = e.target.closest('.btn-delete-asset');
-        if (!deleteButton) return;
-
-        const assetIdToDelete = deleteButton.dataset.id;
-        if (!assetIdToDelete) return;
-
-        // Llamar a la función confirmDestroy con la función de eliminación
-        confirmDestroy(async function() {
-            try {
-                const response = await fetch(`/assets/${assetIdToDelete}`, {
-                    method: 'DELETE',
-                    headers: {
-                        'X-CSRF-TOKEN': csrfToken
-                    }
-                });
-
-                const result = await response.json();
-
-                if (!result.ok) {
-                    showAlert(result.message || 'Error al eliminar el activo', "red", "Error");
-                    return;
-                }
-
-                showAlert(result.message || 'Activo eliminado correctamente.', "green", "Éxito", () => {
-                    const table = $('#file_export').DataTable();
-                    const row = document.querySelector(`.btn-delete-asset[data-id="${assetIdToDelete}"]`).closest('tr');
-                    table.row(row).remove().draw();
-                });
-
-            } catch (error) {
-                showAlert('Ocurrió un error al eliminar el activo: ' + error.message, "red", "Error inesperado");
-            }
-        });
-    });
-});
